Reorganize accelerated training loop to work as mixin
