+++ src/pykeen/training/__init__.py
@@ -12,30 +12,25 @@ slcwa   :class:`pykeen.training.SLCWATrainingLoop`
 .. note:: This table can be re-generated with ``pykeen ls trainers -f rst``
 """
 
-from typing import Set, Type
-
 from class_resolver import Resolver
 
-from .lcwa import LCWATrainingLoop  # noqa: F401
-from .slcwa import SLCWATrainingLoop  # noqa: F401
-from .training_loop import NonFiniteLossError, TrainingLoop  # noqa: F401
+from .lcwa import AcceleratedLCWATrainingLoop, LCWATrainingLoop  # noqa: F401
+from .slcwa import AcceleratedSLCWATrainingLoop, SLCWATrainingLoop  # noqa: F401
+from .training_loop import AcceleratedTrainingLoop, NonFiniteLossError, TrainingLoop  # noqa: F401
 
 __all__ = [
     'TrainingLoop',
     'SLCWATrainingLoop',
     'LCWATrainingLoop',
+    'AcceleratedSLCWATrainingLoop',
+    'AcceleratedLCWATrainingLoop',
     'NonFiniteLossError',
     'training_loop_resolver',
 ]
 
 _TRAINING_LOOP_SUFFIX = 'TrainingLoop'
-_TRAINING_LOOPS: Set[Type[TrainingLoop]] = {
-    LCWATrainingLoop,
-    SLCWATrainingLoop,
-}
-training_loop_resolver = Resolver(
-    _TRAINING_LOOPS,
-    base=TrainingLoop,  # type: ignore
+training_loop_resolver = Resolver.from_subclasses(
+    base=TrainingLoop,
     default=SLCWATrainingLoop,
-    suffix=_TRAINING_LOOP_SUFFIX,
+    skip={AcceleratedTrainingLoop},
 )
